Refactor out rules.py
