+++ src/apologies/character.py
@@ -2,51 +2,16 @@
 # vim: set ft=python ts=4 sw=4 expandtab:
 
 """
-Characters and the actions they can take.
+Player character interfaces.
 """
 
-from enum import Enum
 from abc import ABC, abstractmethod
-from typing import Sequence, Optional
+from typing import Optional
 
 import attr
 
-from .game import Game, GameMode, Player, Card, Pawn
-
-
-class ActionType(Enum):
-    """Enumeration of all actions that a character can take."""
-
-    MOVE_FROM_START = "Move from start"  # Move a pawn from start
-    MOVE_FORWARD = "Move forward"  # Move a pawn forward a certain number of spaces
-    MOVE_BACKARD = "Move backward"  # Move a pawn backward a certain number of spaces
-    CHANGE_PLACES = "Change places"  # Change places, swapping two pawns on the board
-    BUMP_TO_START = "Bump to start"  # Move a pawn from start, bumping another pawn off the board
-
-
-@attr.s
-class Action:
-    """An action that can be taken as part of a move."""
-
-    actiontype = attr.ib(type=ActionType)
-    mine = attr.ib(type=Pawn)
-    theirs = attr.ib(default=None, type=Optional[Pawn])
-    squares = attr.ib(default=None, type=Optional[int])
-
-
-@attr.s
-class Move:
-
-    """
-    A player's move on the board, which consists of one or more actions
-
-    Attributes:
-        card(Card): the card that is being played by this move
-        actions(:obj: List of :obj: Action): List of actions to execute
-    """
-
-    card = attr.ib(type=Card)
-    actions = attr.ib(type=Sequence[Action])
+from .game import Game, GameMode, Player, Card
+from .rules import Move
 
 
 class CharacterInputSource(ABC):
