+++ tests/test_game.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 # vim: set ft=python ts=4 sw=4 expandtab:
 # pylint: disable=no-self-use,protected-access
-# Unit tests for game.py
-
 
 import arrow
 import pytest
